feat: v0.8.5

+++ src/copy/async/fn.rs
@@ -38,14 +38,14 @@ pub async fn async_copy_file(src: &str, dest: &str) -> Result<(), Error> {
 pub async fn async_copy_dir_files(src_dir: &str, dest_dir: &str) -> Result<(), Error> {
     let src_path: &Path = Path::new(src_dir);
     let dest_path: &Path = Path::new(dest_dir);
-    if dest_path.exists() {
-        if let Some(dest_path_str) = dest_path.to_str() {
-            if dest_path.is_file() {
-                delete_file(dest_path_str)?;
-            }
-            if dest_path.is_dir() {
-                delete_dir(dest_path_str)?;
-            }
+    if dest_path.exists()
+        && let Some(dest_path_str) = dest_path.to_str()
+    {
+        if dest_path.is_file() {
+            delete_file(dest_path_str)?;
+        }
+        if dest_path.is_dir() {
+            delete_dir(dest_path_str)?;
         }
     }
     create_dir_all(dest_path).await?;
